+++ tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -1,9 +1,8 @@
-package org.apache;
+package org.apache.coyote.http11.controller;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UnauthorizedException;
 import nextstep.jwp.model.User;
-import org.apache.coyote.AbstractController;
 import org.apache.coyote.http11.exception.FileNotReadableException;
 import org.apache.coyote.http11.message.ContentType;
 import org.apache.coyote.http11.message.HttpStatus;
